take into account distance to guard for vertexScore and finished merge
