--- src/main/java/app/model/agents/Evasion/EvasionStrategy.java
@@ -1,9 +1,0 @@
-package app.model.agents.Evasion;
-
-public enum EvasionStrategy
-{
-    RANDOM,
-    RANDOMDIRECTED,
-    DIRECTED,
-    INTELLIGENT
-}
